+++ src/features/blog-feed/index.astro
@@ -1,5 +1,5 @@
 ---
-import BlogFeedContainer from './blog-feed-ui';
+import BlogFeedContainer from './ui';
 
 type Post = {
   title: string;
