+++ src/features/blog-feed/ui.tsx
@@ -15,7 +15,7 @@ import {
   TableRow,
 } from '@stylelist94/nine-beauty-actress';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 type Props = {
   postList: {
