+++ src/features/color-picker/index.ts
@@ -1,0 +1,1 @@
+export { default as ColorPicker } from './ui';
