+++ src/features/icon-converter/index.ts
@@ -1,0 +1,2 @@
+export { default as IconConverter } from './ui';
+export type * from './types';
