+++ src/features/landing-hero/index.ts
@@ -1,0 +1,1 @@
+export { default as TopContent } from './ui';
