+++ src/features/play-ball/index.astro
@@ -1,5 +1,5 @@
 ---
-import PlayBallContainer from './play-ball-ui';
+import PlayBallContainer from './ui';
 
 type TableCellData = {
   Text: string;
