+++ src/features/play-ball/ui.tsx
@@ -20,7 +20,7 @@ import {
   ScrollBar,
 } from '@stylelist94/nine-beauty-actress';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 
 type TeamStanding = {
