+++ src/features/search-explore/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchExplore } from './ui';
