+++ src/features/spring-simulator/index.ts
@@ -1,0 +1,1 @@
+export { default as SpringSimulator } from './ui';
