+++ src/features/spring-simulator/ui.tsx
@@ -24,7 +24,7 @@ import {
   Combobox,
 } from '@stylelist94/nine-beauty-actress';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 type PresetOptionValue = 'scroll-linked';
