+++ src/features/string-maker/index.ts
@@ -1,0 +1,1 @@
+export { default as StringMaker } from './ui';
