+++ src/layouts/root.astro
@@ -1,9 +1,9 @@
 ---
-import Head from '@/components/head.astro';
-import Toaster from '@/components/ui/sonner';
-import ThemeScript from '@/components/theme-script.astro';
+import Head from '@/shared/ui/head.astro';
+import Toaster from '@/shared/ui/sonner';
+import ThemeScript from '@/shared/ui/theme-script.astro';
 
-import '@/styles/global.css';
+import '@/shared/styles/global.css';
 ---
 
 <ThemeScript />
