+++ src/pages/productivity.astro
@@ -3,23 +3,23 @@ export const prerender = false;
 
 import { MainContainer } from '@stylelist94/nine-beauty-actress';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 import RootLayout from '@/layouts/root.astro';
-import Header from '@/components/header';
-import PageFooter from '@/components/page-footer.astro';
+import Header from '@/shared/ui/header';
+import PageFooter from '@/shared/ui/page-footer.astro';
 
-import ColorPicker from '@/components/color-picker';
-import SearchExplore from '@/components/search-explore';
-import StringMaker from '@/components/string-maker';
-import SpringSimulator from '@/components/spring-simulator';
-import TwBreakpoint from '@/components/tw-breakpoint';
-import UnitConverter from '@/components/unit-converter';
-import IconConverter from '@/components/icon-converter';
-import BlogFeed from '@/components/blog-feed.astro';
-import PlayBall from '@/components/play-ball.astro';
+import { ColorPicker } from '@/features/color-picker';
+import { SearchExplore } from '@/features/search-explore';
+import { StringMaker } from '@/features/string-maker';
+import { SpringSimulator } from '@/features/spring-simulator';
+import { TwBreakpoint } from '@/features/tw-breakpoint';
+import { UnitConverter } from '@/features/unit-converter';
+import { IconConverter } from '@/features/icon-converter';
+import BlogFeed from '@/features/blog-feed/index.astro';
+import PlayBall from '@/features/play-ball/index.astro';
 
-import '@/styles/global.css';
+import '@/shared/styles/global.css';
 ---
 
 <RootLayout>
