+++ src/shared/ui/page-footer.astro
@@ -1,9 +1,9 @@
 ---
 import { Footer } from '@stylelist94/nine-beauty-actress';
 
-import { cn } from '@/lib/utils';
+import { cn } from '../lib/utils';
 
-import ThemeControlSwitch from '@/components/theme-control-switch';
+import ThemeControlSwitch from './theme-control-switch';
 ---
 
 <Footer>
