+++ src/shared/ui/theme-script.astro
