+++ src/shared/ui/typewriter.tsx
